refactor: rename component files
