+++ src/app/components/RecipesListing.tsx
@@ -3,7 +3,7 @@
 import styled from "styled-components";
 
 import type { Recipe } from "../lib/types";
-import { RecipeItem } from "./recipeItem";
+import { RecipeItem } from "./RecipeItem";
 
 type RecipesProps = {
   recipes: Recipe[];
